Fix various layout problems by simplifying data
While deploying AngularGridster for our customers, I had numerous
reports of funny corner cases where the layout would wind up being
incorrect. After much debugging, I found that the interaction of
the `grid` data structure with the various movements my users
reported problems with meant that the data structure simply wasn't
able to do the task they wanted. For example, dragging an item
over another from the bottom would always cause the stationary item
to be hidden under the moving one if: 1) the moving item were at
least as large as the stationary one, and 2) if the moving item's
column was the same as the stationary one.

The problem results because the `grid` data structure doesn't permit
mutliple items in the same grid cell. When performing the maneuver
described above, this is briefly necessary in order to properly
detect the collisions of the various items. Since the data structure
makes this impossible, the task fails.

To address these problems, I converted the `grid` into a simple
array of items  and updated all the functions which use it. This
allows there to be multiple items at the same position, and allows
the algorithms to straighten things out in an orderly fashion.