Automatically scroll while dragging
If a user drags an item to the top or bottom of the page, they must
continually jiggle the item to keep the scrolling going as scrolling
only happens when mouse events are created. If no new mouse events
are created, then scrolling stops even if the item is still within
the scrolling zone at the top or bottom of the page.

This change sets a timer so that as long as the user keeps the
mouse in the scroll zone, simulated mouse events will be generated,
thus causing the scrolling to continue.